refactor: Update almost all files
