+++ setup.cfg
@@ -1,9 +1,9 @@
 [metadata]
-name = project-template
-version = 0.0.0
+name = unet-pytorch
+version = 0.0.1
 author = Yakhyokhuja Valikhujaev
 author_email = yakhyo9696@gmail.com
-description = Template for ML/DL projects
+description = Semantic Segmentation using UNet
 readme = README.md
 python_requires = >=3.8.0
 license_file = LICENSE
@@ -18,10 +18,8 @@ max-line-length = 120
 [flake8]
 # note: we ignore all 501s (line too long) anyway as they're taken care of by black
 max-line-length = 120
-ignore = E203, E402, W503, W504, F821, E501, E266, F401, F841, F541
 per-file-ignores =
     __init__.py: F401, F403, F405
-    ./hubconf.py: F401
 exclude = docs*
 
 [pydocstyle]
