+++ tools/download.sh
@@ -1,0 +1,33 @@
+!/bin/bash
+
+if [[ ! -f ~/.kaggle/kaggle.json ]]; then
+  echo -n "Kaggle username: "
+  read USERNAME
+  echo
+  echo -n "Kaggle API key: "
+  read APIKEY
+
+  mkdir -p ~/.kaggle
+  echo "{\"username\":\"$USERNAME\",\"key\":\"$APIKEY\"}" > ~/.kaggle/kaggle.json
+  chmod 600 ~/.kaggle/kaggle.json
+fi
+
+pip install kaggle --upgrade
+
+# creating folders
+mkdir data
+mkdir data/images
+mkdir data/masks
+
+# downloading images and masks then moving to the created forlders
+kaggle competitions download -c carvana-image-masking-challenge -f train_hq.zip
+unzip train_hq.zip
+mv train_hq/* data/images/
+rm -d train_hq
+rm train_hq.zip
+
+kaggle competitions download -c carvana-image-masking-challenge -f train_masks.zip
+unzip train_masks.zip
+mv train_masks/* data/masks/
+rm -d train_masks
+rm train_masks.zip
